requirements.txt
python-telegram-bot 
requests 
playwright